// src/types/review.d.ts
export interface Review {
  review_title: string;
  review_id: number;
  review_content: string;
  like_count: number;
  course_name: string;
  voted: boolean;
  instructor_name: string;
  taken_semyr: string;
  grade: string;
  liked: boolean;
  review_time: string;
  mine: boolean;
  updated: boolean;
}